Fix: Fix the issue with changing the category when editing real estate and update the functionality for deleting real estate in the admin panel

+++ batdongsan/src/main/java/batdongsan/controllers/admin/RealEstateController.java
@@ -110,7 +110,13 @@ public class RealEstateController {
 			query.setParameter("realEstateId", realEstateId);
 			HCMRealEstateModel deletedRealEstate = query.uniqueResult();
 			
-			session.delete(deletedRealEstate);
+			if  ( deletedRealEstate.getInterestedClick() <= 0 ) {
+				session.delete(deletedRealEstate);
+			}
+			else {
+				deletedRealEstate.setDeleteStatus(true);
+				session.update(deletedRealEstate);
+			}
 			t.commit();
 			return "redirect:/admin/quan-ly-bat-dong-san.html";
 		} catch (Exception e) {
